+++ src/lang-xpath/test/uk/co/reecedunn/intellij/plugin/xpath/tests/lexer/XPathLexerTest.kt
@@ -1364,4 +1364,15 @@ class XPathLexerTest : LexerTestCase() {
         matchSingleToken(lexer, ",", XPathTokenType.COMMA)
         matchSingleToken(lexer, ")", XPathTokenType.PARENTHESIS_CLOSE)
     }
+
+    @Test
+    @DisplayName("XPath 4.0 ED EBNF (128) EnumerationType")
+    fun enumerationType() {
+        val lexer = createLexer()
+
+        matchSingleToken(lexer, "enum", XPathTokenType.K_ENUM)
+        matchSingleToken(lexer, "(", XPathTokenType.PARENTHESIS_OPEN)
+        matchSingleToken(lexer, ",", XPathTokenType.COMMA)
+        matchSingleToken(lexer, ")", XPathTokenType.PARENTHESIS_CLOSE)
+    }
 }