+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/intellij/lexer/XQuerySyntaxHighlighterTest.kt
@@ -355,6 +355,9 @@ class XQuerySyntaxHighlighterTest {
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_ENTIRE).size, `is`(1))
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_ENTIRE)[0], `is`(XQuerySyntaxHighlighterColors.KEYWORD))
 
+        assertThat(highlighter.getTokenHighlights(XPathTokenType.K_ENUM).size, `is`(1))
+        assertThat(highlighter.getTokenHighlights(XPathTokenType.K_ENUM)[0], `is`(XQuerySyntaxHighlighterColors.KEYWORD))
+
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_EQ).size, `is`(1))
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_EQ)[0], `is`(XQuerySyntaxHighlighterColors.KEYWORD))